Add metadata-only digital objects, refs #9969
This commit adds a new type of digital object (usage_id) meant to be used
when AtoM doesn't hold the digital object but it keeps its metadata. We've
called this new type or usage: `Offline`.

Some componentes like OAI, EAD export, CLI tasks or the ES indexer learned how
to operate with this new type.

The HTTP API has also been updated to accept some extra fields that are relevant
in metadata-only uploads: format name, format version, format registry key and
format registry name.

+++ lib/model/QubitDigitalObject.php
@@ -1243,14 +1243,17 @@ class QubitDigitalObject extends BaseDigitalObject
       $child->delete();
     }
 
-    // Delete digital asset
-    if (file_exists($this->getAbsolutePath()))
+    if ($this->usageId !== QubitTerm::OFFLINE_ID)
     {
-      unlink($this->getAbsolutePath());
-    }
+      // Delete digital asset
+      if (file_exists($this->getAbsolutePath()))
+      {
+        unlink($this->getAbsolutePath());
+      }
 
-    // Prune asset directory, if empty
-    self::pruneEmptyDirs(sfConfig::get('sf_web_dir').$this->path);
+      // Prune asset directory, if empty
+      self::pruneEmptyDirs(sfConfig::get('sf_web_dir').$this->path);
+    }
 
     foreach (QubitRelation::getBySubjectOrObjectId($this->id) as $item)
     {
@@ -1577,9 +1580,12 @@ class QubitDigitalObject extends BaseDigitalObject
    */
   public function getPublicPath()
   {
-    if ($this->usageId == QubitTerm::EXTERNAL_URI_ID) {
+    if ($this->usageId == QubitTerm::EXTERNAL_URI_ID)
+    {
       return $this->getPath();
-    } else {
+    }
+    else
+    {
       return public_path($this->getFullPath(), true);
     }
   }
